requirements.txt
# Generated via pip freeze
alembic==1.13.2
asyncpg==0.29.0
bcrypt==4.1.3
certifi==2024.7.4
charset-normalizer==3.3.2
chromadb==0.5.4
click==8.1.7
fastapi==0.111.1
fastapi-cache2[redis]==0.2.2
google-auth==2.32.0
h11==0.14.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
onnxruntime==1.18.1
openai==1.37.1
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
posthog==3.5.0
protobuf==5.27.2
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
schedule==1.2.2
sentence-transformers==3.0.1
six==1.16.0
slowapi==0.1.9
SQLAlchemy==2.0.31
structlog==24.3.0
twilio==9.2.2
typing_extensions==4.12.2
uvicorn==0.30.3
urllib3==2.2.2

# ========================================
# NEXT-GEN AI CONTENT EMPIRE ENHANCEMENTS
# ========================================

# Advanced AI Models & Providers
anthropic>=0.34.0
google-generativeai>=0.8.0
ollama>=0.3.0
transformers>=4.45.0
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
diffusers>=0.30.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
xformers>=0.0.22

# Specialized AI Services
elevenlabs>=1.0.0
stability-sdk>=0.8.0
replicate>=0.25.0
runpod>=1.5.0
together>=1.0.0
cohere>=5.0.0
mistralai>=1.0.0

# Video & Audio Processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0
pydub>=0.25.1
librosa>=0.10.1
opencv-python>=4.8.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Advanced Image Generation & Processing
controlnet-aux>=0.4.0
insightface>=0.7.3
facexlib>=0.3.0
gfpgan>=1.3.8
realesrgan>=0.3.0

# Social Media APIs & Automation
instagrapi>=2.0.0
tiktok-api>=6.0.0
youtube-transcript-api>=0.6.0
tweepy>=4.14.0
linkedin-api>=2.0.0

# Advanced Analytics & Intelligence
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
dash>=2.14.0

# Web Scraping & Data Collection
scrapy>=2.11.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
playwright>=1.40.0

# Enterprise Features
redis>=5.0.0
celery>=5.3.0
flower>=2.0.0
prometheus-client>=0.19.0
grafana-api>=1.0.3

# Security & Compliance
cryptography>=41.0.0
passlib>=1.7.4
python-jose>=3.3.0
bleach>=6.1.0